+++ cli/src/core/index.ts
@@ -186,6 +186,9 @@ export async function runCli(): Promise<CliResults> {
       case "prisma":
         cliResults.packages.push("prisma")
         break
+      case "drizzle":
+        cliResults.packages.push("drizzle")
+        break
       default:
         break
     }
@@ -287,6 +290,7 @@ export async function runCli(): Promise<CliResults> {
         choices: [
           { value: "none", name: "None" },
           { value: "prisma", name: "Prisma" },
+          { value: "drizzle", name: "Drizzle" },
         ],
         default: !defaultOptions.flags.orm,
       })
@@ -352,6 +356,9 @@ export async function runCli(): Promise<CliResults> {
     switch (project.orm) {
       case "prisma":
         packages.push("prisma")
+        break
+      case "drizzle":
+        packages.push("drizzle")
         break
       default:
         break