+++ cli/src/installers/index.ts
@@ -5,6 +5,7 @@ import { type PackageManager } from "@/utils/get-user-pkg-manager.js"
 
 import { authjsInstaller } from "./authjs.js"
 import { biomeInstaller } from "./biome.js"
+import { drizzleInstaller } from "./drizzle.js"
 import { envVariablesInstaller } from "./env-vars.js"
 import { eslintInstaller } from "./eslint.js"
 import { prismaInstaller } from "./prisma.js"
@@ -15,6 +16,7 @@ export const availablePackages = [
   "authjs",
   "betterAuth",
   "prisma",
+  "drizzle",
   "envVariables",
   "eslint/prettier",
   "biome",
@@ -29,7 +31,7 @@ export type BackendFramework = (typeof backendFrameworks)[number]
 export const authProviders = ["none", "authjs", "betterAuth"] as const
 export type AuthProvider = (typeof authProviders)[number]
 
-export const databaseORM = ["none", "prisma"] as const
+export const databaseORM = ["none", "prisma", "drizzle"] as const
 export type DatabaseORM = (typeof databaseORM)[number]
 
 export const databaseProviders = ["sqlite", "mysql", "postgresql"] as const
@@ -71,6 +73,10 @@ export const buildPkgInstallerMap = (
   prisma: {
     inUse: packages.includes("prisma"),
     installer: prismaInstaller,
+  },
+  drizzle: {
+    inUse: packages.includes("drizzle"),
+    installer: drizzleInstaller,
   },
   envVariables: {
     inUse: true,