+++ cli/CHANGELOG.md
@@ -1,5 +1,21 @@
 # create-lx2-app
 
+## 0.9.0
+
+### Minor Changes
+
+- [#154](https://github.com/SlickYeet/create-lx2-app/pull/154)
+  [`0e2f227`](https://github.com/SlickYeet/create-lx2-app/commit/0e2f22702ebd149385a7b0f941365c9b3f911aff)
+  Thanks [@SlickYeet](https://github.com/SlickYeet)! - add drizzle to cli
+
+## 0.8.0
+
+### Major Changes
+
+- [#143](https://github.com/SlickYeet/create-lx2-app/pull/143)
+  [`ae1be44`](https://github.com/SlickYeet/create-lx2-app/commit/ae1be44652a99d2dd34d75e460995a14091c6cfd)
+  Thanks [@SlickYeet](https://github.com/SlickYeet)! - add drizzle to cli
+
 ## 0.7.2
 
 ### Patch Changes